+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix output of binary data using hex-encoding
+- Print proper Python boolean literals for json requests
 
 ## [1.1.0] - 2023-01-16